requirements.txt
streamlit
langchain
langchain-community
langchain-experimental
langchain-google-genai
mysql-connector-python
python-dotenv
sqlalchemy